Add KaTeX, remove AWS Lambda functions
Since rendering markdown into HTML is done locally now, and this file is
no longer run as a Lambda function, the handler function and the AWS
functions it references can be removed.

This also starts the migration from bespoke math rendering to using
KaTex, with the room-area blog entry as the first conversion.

I'm also adding some comments to the marked override functions showing
what they're for, with examples.