Use handlebars, generate and publish whole site
This adds handlebars.js templating to help with some complex rendering
logic:
- Whether about or blog is highlighted
- When to include KaTeX
- Nicer OG tag handling

This replaces test.js/index.js for generating one file at a time with
index.js/render.js. The new index.js takes an optional file glob param,
but defaults to 'src/**/*.md'. It iterates over each file in the pattern
and renders it as an HTML file in the dist directory.

Lastly, this adds a publish.zsh script to upload all the HTML files to
an S3 bucket.